:recycle: Move useBlockchainApplicationsManagement to single file
