+++ src/modules/BlockchainApplication/components/ApplicationSwitcher/index.js
@@ -4,7 +4,7 @@ import { useTheme } from 'hooks/useTheme';
 import { P } from 'components/shared/toolBox/typography';
 import ChangeSvg from 'assets/svgs/ChangeSvg';
 import getStyles from './styles';
-import { useBlockchainApplicationsManagement } from '../../context/BlockchainApplicationsManagementContext';
+import { useBlockchainApplicationsManagement } from '../../hooks/useBlockchainApplicationManagement';
 
 const ApplicationSwitcher = ({ onPress }) => {
   const { styles } = useTheme({ styles: getStyles });
