+++ src/utilities/api/hooks/useQueryKeys.js
@@ -1,5 +1,5 @@
 // eslint-disable-next-line import/no-cycle
-import { useBlockchainApplicationsManagement } from 'modules/BlockchainApplication/context/BlockchainApplicationsManagementContext';
+import { useBlockchainApplicationsManagement } from 'modules/BlockchainApplication/hooks/useBlockchainApplicationManagement';
 import { METHOD } from 'utilities/api/constants';
 import { APPLICATION } from '../queries';
